管理员授权功能(RBAC) version 1.10

+++ backend/controllers/PostController.php
@@ -9,6 +9,8 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use common\models\Tag;
+use yii\filters\AccessControl;
+use yii\web\ForbiddenHttpException;
 
 /**
  * PostController implements the CRUD actions for Post model.
@@ -27,6 +29,21 @@ class PostController extends Controller
                     'delete' => ['POST'],
                 ],
             ],
+        	'access' => [
+        			'class' => AccessControl::className(),
+        			'rules' => [
+        					[
+        							'actions' => ['index', 'view'],
+        							'allow' => true,
+        							'roles' => ['?'],
+        					],
+        					[
+        							'actions' => ['index', 'view', 'create', 'update'],
+        							'allow' => true,
+        							'roles' => ['@'],
+        					],
+        			],
+        	],
         ];
     }
 
@@ -77,6 +94,10 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+    	if (!\Yii::$app->user->can('createPost')){
+    		throw new ForbiddenHttpException('对不起，你没有进行该操作的权限！');
+    	}
+    	
         $model = new Post();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -96,6 +117,10 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+    	if (!\Yii::$app->user->can('updatePost')){
+    		throw new ForbiddenHttpException('对不起，你没有进行该操作的权限！');
+    	}
+    	
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -115,6 +140,10 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+    	if (!\Yii::$app->user->can('deletePost')){
+    		throw new ForbiddenHttpException('对不起，你没有进行该操作的权限！');
+    	}
+    	 
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
